refactor: RestClient를 Wrapping하는 KakaoClient 생성, 함수 리팩토링
